Refactor PublicKeyCredentialCreationOptions
