+++ app/src/test/java/com/x8bit/bitwarden/ui/vault/feature/addedit/util/PublicKeyCredentialCreationOptionsTestHelpers.kt
@@ -1,37 +1,37 @@
 package com.x8bit.bitwarden.ui.vault.feature.addedit.util
 
-import com.x8bit.bitwarden.data.autofill.fido2.datasource.network.model.PublicKeyCredentialCreationOptions
+import com.x8bit.bitwarden.data.autofill.fido2.model.PasskeyAttestationOptions
 
 /**
- * Returns a mock FIDO 2 [PublicKeyCredentialCreationOptions] object to simulate a credential
+ * Returns a mock FIDO 2 [PasskeyAttestationOptions] object to simulate a credential
  * creation request.
  */
 @Suppress("MaxLineLength")
 fun createMockPublicKeyCredentialCreationOptions(
     number: Int,
-    userVerificationRequirement: PublicKeyCredentialCreationOptions.AuthenticatorSelectionCriteria.UserVerificationRequirement? = null,
-) = PublicKeyCredentialCreationOptions(
-    authenticatorSelection = PublicKeyCredentialCreationOptions
+    userVerificationRequirement: PasskeyAttestationOptions.AuthenticatorSelectionCriteria.UserVerificationRequirement? = null,
+) = PasskeyAttestationOptions(
+    authenticatorSelection = PasskeyAttestationOptions
         .AuthenticatorSelectionCriteria(userVerification = userVerificationRequirement),
     challenge = "mockPublicKeyCredentialCreationOptionsChallenge-$number",
     excludeCredentials = listOf(
-        PublicKeyCredentialCreationOptions.PublicKeyCredentialDescriptor(
+        PasskeyAttestationOptions.PublicKeyCredentialDescriptor(
             type = "mockPublicKeyCredentialDescriptorType-$number",
             id = "mockPublicKeyCredentialDescriptorId-$number",
             transports = listOf("mockPublicKeyCredentialDescriptorTransports-$number"),
         ),
     ),
     pubKeyCredParams = listOf(
-        PublicKeyCredentialCreationOptions.PublicKeyCredentialParameters(
+        PasskeyAttestationOptions.PublicKeyCredentialParameters(
             type = "PublicKeyCredentialParametersType-$number",
             alg = number.toLong(),
         ),
     ),
-    relyingParty = PublicKeyCredentialCreationOptions.PublicKeyCredentialRpEntity(
+    relyingParty = PasskeyAttestationOptions.PublicKeyCredentialRpEntity(
         name = "mockPublicKeyCredentialRpEntityName-$number",
         id = "mockPublicKeyCredentialRpEntity-$number",
     ),
-    user = PublicKeyCredentialCreationOptions.PublicKeyCredentialUserEntity(
+    user = PasskeyAttestationOptions.PublicKeyCredentialUserEntity(
         name = "mockPublicKeyCredentialUserEntityName-$number",
         id = "mockPublicKeyCredentialUserEntityId-$number",
         displayName = "mockPublicKeyCredentialUserEntityDisplayName-$number",
